Crate flux_refineck

Refinement type checking

Modules

checker 🔒

compare_impl_item

errors 🔒

ghost_statements 🔒

Ghost statements are statements that are not part of the original mir, but are added from information extracted from the compiler or some additional analysis.

invariants

primops 🔒

queue 🔒

type_env 🔒

Functions

add_fn_fix_diagnostic 🔒

call_error 🔒

check_body 🔒

check_fn

check_static

fn_first_line 🔒

report_errors 🔒

report_expected_neg 🔒

report_fixpoint_errors

rerun_hint_note 🔒

ret_error 🔒

shell_quote_arg 🔒