Enum flux_refineck::fixpoint_encoding::fixpoint::Var

pub enum Var {
    Underscore,
    Global(GlobalVar),
    Local(LocalVar),
    TupleCtor {
        arity: usize,
    },
    TupleProj {
        arity: usize,
        field: u32,
    },
    UIFRel(BinRel),
    Itf(Symbol),
    ConstGeneric(ParamConst),
}

Variants

Underscore

Global(GlobalVar)

Local(LocalVar)

TupleCtor

Fields

arity: usize

TupleProj

Fields

arity: usize

field: u32

UIFRel(BinRel)

Itf(Symbol)

Interpreted theory function. This can be an arbitrary string, thus we are assuming the name is different than the display implementation for the other variants.

ConstGeneric(ParamConst)