Struct flux_refineck::fixpoint_encoding::ExprEncodingCtxt

struct ExprEncodingCtxt<'genv, 'tcx> {
    genv: GlobalEnv<'genv, 'tcx>,
    global_var_gen: IndexGen<GlobalVar>,
    const_map: FxIndexMap<Key<'tcx>, ConstInfo>,
    dbg_span: Span,
}

Fields
genv: GlobalEnv<'genv, 'tcx>global_var_gen: IndexGen<GlobalVar>const_map: FxIndexMap<Key<'tcx>, ConstInfo>dbg_span: Span

Used to report bugs

Implementations

impl<'genv, 'tcx> ExprEncodingCtxt<'genv, 'tcx>

fn new(genv: GlobalEnv<'genv, 'tcx>, dbg_span: Span) -> QueryResult<Self>

fn expr_to_fixpoint( &mut self, expr: &Expr, env: &Env, ) -> QueryResult<Expr<FixpointTypes>>

fn exprs_to_fixpoint<'b>( &mut self, exprs: impl IntoIterator<Item = &'b Expr>, env: &Env, ) -> QueryResult<Vec<Expr<FixpointTypes>>>

fn un_op_to_fixpoint( &mut self, op: UnOp, e: &Expr, env: &Env, ) -> QueryResult<Expr<FixpointTypes>>

fn bin_op_to_fixpoint( &mut self, op: &BinOp, e1: &Expr, e2: &Expr, env: &Env, ) -> QueryResult<Expr<FixpointTypes>>

fn bin_rel_to_fixpoint( &mut self, sort: &Sort, rel: BinRel, e1: &Expr, e2: &Expr, env: &Env, ) -> QueryResult<Expr<FixpointTypes>>

A binary relation is encoded as a structurally recursive relation between aggregate sorts. For “leaf” expressions, we encode them as an interpreted relation if the sort supports it, otherwise we use an uninterpreted function. For example, consider the following relation between two tuples of sort (int, int -> int)

(0, λv. v + 1) <= (1, λv. v + 1)

The encoding in fixpoint will be

0 <= 1 && (le (λv. v + 1) (λv. v + 1))

Where <= is the (interpreted) less than or equal relation between integers and le is an uninterpreted relation between (the encoding of) lambdas.

fn apply_bin_rel_rec( &mut self, sorts: &[Sort], rel: BinRel, e1: &Expr, e2: &Expr, env: &Env, mk_proj: impl Fn(u32) -> FieldProj, ) -> QueryResult<Expr<FixpointTypes>>

Apply binary relation recursively over aggregate expessions

fn func_to_fixpoint(&mut self, func: &Expr, env: &Env) -> Var

fn imm( &mut self, arg: &Expr, sort: &Sort, env: &mut Env, bindings: &mut Vec<(LocalVar, Sort<FixpointTypes>, Expr<FixpointTypes>)>, ) -> QueryResult<Var>

fn register_const_for_alias_reft( &mut self, alias_reft: &AliasReft, arity: usize, ) -> GlobalVar

returns the ‘constant’ UIF for Var used to represent the alias_pred, creating and adding it to the const_map if necessary

fn register_const_for_lambda(&mut self, lam: &Lambda) -> GlobalVar

We encode lambdas with uninterpreted constant. Two syntactically equal lambdas will be encoded with the same constant.

fn qualifier_to_fixpoint( &mut self, qualifier: &Qualifier, ) -> QueryResult<Qualifier<FixpointTypes>>