Struct flux_refineck::fixpoint_encoding::FixpointCtxt

pub struct FixpointCtxt<'genv, 'tcx, T: Eq + Hash> {
    comments: Vec<String>,
    genv: GlobalEnv<'genv, 'tcx>,
    kvars: KVarStore,
    sorts: SortStore,
    kcx: KVarEncodingCtxt,
    ecx: ExprEncodingCtxt<'genv, 'tcx>,
    env: Env,
    tags: IndexVec<TagIdx, T>,
    tags_inv: UnordMap<T, TagIdx>,
    def_id: LocalDefId,
}

Fields
comments: Vec<String>genv: GlobalEnv<'genv, 'tcx>kvars: KVarStoresorts: SortStorekcx: KVarEncodingCtxtecx: ExprEncodingCtxt<'genv, 'tcx>env: Envtags: IndexVec<TagIdx, T>tags_inv: UnordMap<T, TagIdx>def_id: LocalDefId

DefId of the item being checked. This can be a function/method or an adt when checking invariants.

Implementations

impl<'genv, 'tcx, Tag> FixpointCtxt<'genv, 'tcx, Tag>
where Tag: Hash + Eq + Copy,

pub fn new( genv: GlobalEnv<'genv, 'tcx>, def_id: LocalDefId, kvars: KVarStore, ) -> Self

pub(crate) fn collect_sorts<T: TypeVisitable>(&mut self, t: &T) -> QueryResult

Collect all the sorts that need to be defined in fixpoint to encode t

pub(crate) fn with_name_map<R>( &mut self, name: Name, f: impl FnOnce(&mut Self, LocalVar) -> R, ) -> R

fn assume_const_val( cstr: Constraint<FixpointTypes>, var: GlobalVar, const_val: Constant, ) -> Constraint<FixpointTypes>

fn fixpoint_const_info(const_info: ConstInfo) -> ConstInfo<FixpointTypes>

pub fn check( self, cache: &mut QueryCache, constraint: Constraint<FixpointTypes>, config: &CheckerConfig, ) -> QueryResult<Vec<Tag>>

fn tag_idx(&mut self, tag: Tag) -> TagIdx
where Tag: Debug,

pub fn head_to_fixpoint( &mut self, expr: &Expr, mk_tag: impl Fn(Option<ESpan>) -> Tag + Copy, ) -> QueryResult<Constraint<FixpointTypes>>
where Tag: Debug,

Encodes an expression in head position as a constraint “peeling out” implications and foralls.

pub fn assumption_to_fixpoint( &mut self, pred: &Expr, ) -> QueryResult<(Vec<(LocalVar, Sort<FixpointTypes>, Expr<FixpointTypes>)>, Pred<FixpointTypes>)>

Encodes an expression in assumptive position as a fixpoint::Pred. Returns the encoded predicate and a possible list of bindings produced by ANF-fing kvars.

fn assumption_to_fixpoint_aux( &mut self, expr: &Expr, bindings: &mut Vec<(LocalVar, Sort<FixpointTypes>, Expr<FixpointTypes>)>, preds: &mut Vec<Pred<FixpointTypes>>, ) -> QueryResult

Auxiliary function to avoid creating nested conjunctions

fn kvar_to_fixpoint( &mut self, kvar: &KVar, bindings: &mut Vec<(LocalVar, Sort<FixpointTypes>, Expr<FixpointTypes>)>, ) -> QueryResult<Pred<FixpointTypes>>

fn def_span(&self) -> Span