flux_refineck

Module ghost_statements

Ghost statements are statements that are not part of the original mir, but are added from information extracted from the compiler or some additional analysis.

Modules

fold_unfold 🔒

points_to 🔒

This module implements a points-to analysis for mutable references.

Structs

GhostStatements 🔒

StatementsAt 🔒

Enums

CheckerId

A type to indicate who the ghost statements are for: either a regular DefId (including closures) a promoted body.

GhostStatement 🔒

Point 🔒

A point in the control flow graph where ghost statements can be inserted.

Functions

all_nested_bodies 🔒

compute_ghost_statements 🔒

Type Aliases

EdgeMap 🔒

LocationMap 🔒