Struct flux_refineck::refine_tree::RefineCtxt

pub(crate) struct RefineCtxt<'a> {
    tree: &'a mut RefineTree,
    ptr: NodePtr,
}

A refinement context tracks all the refinement parameters and predicates available in a particular path during type-checking. For example, consider the following program:

#[flux::sig(fn(i32[@a0], i32{v : v > a0})) -> i32]
fn add(x: i32, y: i32) -> i32 {
    x + y
}

At the beginning of the function, the refinement context will be {a0: int, a1: int, a1 > a0}, where a1 is a freshly generated name for the existential variable in the refinement of y.

More specifically, a RefineCtxt represents a path from the root to some internal node in a refinement tree.

Fields
tree: &'a mut RefineTreeptr: NodePtr
Implementations

impl<'rcx> RefineCtxt<'rcx>

pub(crate) fn as_subtree(&mut self) -> RefineSubtree<'_>

pub fn subtree_at(&mut self, snapshot: &Snapshot) -> Option<RefineSubtree<'_>>

pub(crate) fn snapshot(&self) -> Snapshot

pub(crate) fn branch(&mut self) -> RefineCtxt<'_>

pub(crate) fn scope(&self) -> Scope

pub(crate) fn push_comment(&mut self, comment: impl ToString) -> RefineCtxt<'_>

pub(crate) fn define_var(&mut self, sort: &Sort) -> Name

Defines a fresh refinement variable with the given sort. It returns the freshly generated name for the variable.

pub(crate) fn define_vars(&mut self, sort: &Sort) -> Expr

Given a sort that may contain nested tuples, it destructs the tuples recursively, generating multiple fresh variables and returning the “eta-expanded” tuple of fresh variables. This is in contrast to generating a single fresh variable of tuple sort.

For example, given the sort (int, (bool, int)) it returns (a0, (a1, a2)) for fresh variables a0: int, a1: bool, and a2: int.

pub(crate) fn assume_pred(&mut self, pred: impl Into<Expr>)

pub(crate) fn check_pred(&mut self, pred: impl Into<Expr>, tag: Tag)

pub(crate) fn check_impl( &mut self, pred1: impl Into<Expr>, pred2: impl Into<Expr>, tag: Tag )

pub(crate) fn unpack( &mut self, ty: &Ty, assume_invariants: AssumeInvariants ) -> Ty

pub(crate) fn unpacker( &mut self, assume_invariants: AssumeInvariants ) -> Unpacker<'_, 'rcx>

pub(crate) fn assume_invariants(&mut self, ty: &Ty, overflow_checking: bool)

pub(crate) fn replace_evars(&mut self, evars: &EVarSol)