Struct flux_refineck::type_env::TypeEnv

pub struct TypeEnv<'a> {
    bindings: PlacesTree,
    local_decls: &'a LocalDecls,
}

Fields
bindings: PlacesTreelocal_decls: &'a LocalDecls
Implementations

impl TypeEnv<'_>

pub fn new(local_decls: &LocalDecls) -> TypeEnv<'_>

pub fn alloc_with_ty(&mut self, local: Local, ty: Ty)

pub fn alloc(&mut self, local: Local)

pub(crate) fn into_infer( self, scope: Scope, ) -> Result<BasicBlockEnvShape, CheckerErrKind>

pub(crate) fn lookup_place( &mut self, rcx: &mut RefineCtxt<'_>, infcx: &InferCtxt<'_, '_, '_>, place: &Place, ) -> Result<Ty, CheckerErrKind>

pub(crate) fn get(&self, path: &Path) -> Ty

pub fn update_path(&mut self, path: &Path, new_ty: Ty)

pub(crate) fn borrow( &mut self, rcx: &mut RefineCtxt<'_>, infcx: &InferCtxt<'_, '_, '_>, re: Region, mutbl: Mutability, place: &Place, ) -> Result<Ty, CheckerErrKind>

When checking a borrow in the right hand side of an assignment x = &'?n p, we use the annotated region '?n in the type of the result. This region will only be used temporarily and then replaced by the region in the type of x after the assignment. See TypeEnv::assign

pub(crate) fn ptr_to_ref_at_place( &mut self, rcx: &mut RefineCtxt<'_>, infcx: &mut InferCtxt<'_, '_, '_>, place: &Place, ) -> Result<(), CheckerErrKind>

pub(crate) fn ptr_to_ref( &mut self, rcx: &mut RefineCtxt<'_>, infcx: &mut InferCtxt<'_, '_, '_>, reason: ConstrReason, re: Region, path: &Path, bound: PtrToRefBound, ) -> Result<Ty, CheckerErrKind>

Convert a (strong) pointer to a mutable reference.

This roughly implements the following inference rule:

                  tโ <: tโ
-------------------------------------------------
ฮโ,โ:t1,ฮโ ; ptr(mut, โ) => ฮโ,โ:โtโ,ฮโ ; &mut t2

The bound tโ can be either inferred (PtrToRefBound::Infer), or explicilty provided (PtrToRefBound::Ty).

As an example, consider the environment x: i32[a] and the pointer ptr(mut, x). Converting the pointer to a mutable reference with an inferred bound produces the following derivation (roughly):

                   i32[a] <: i32{v: $k(v)}
----------------------------------------------------------------
x: i32[a] ; ptr(mut, x) => x:โi32{v: $k(v)} ; &mut i32{v: $k(v)}

pub(crate) fn assign( &mut self, rcx: &mut RefineCtxt<'_>, infcx: &mut InferCtxt<'_, '_, '_>, place: &Place, new_ty: Ty, ) -> Result<(), CheckerErrKind>

Updates the type of place to new_ty

This process involves recovering the original regions (lifetimes) used in the (unrefined) Rust type of place and then substituting these regions in new_ty. For instance, if we are assigning a value of type S<&'?10 i32{v: v > 0}> to a variable x, and the (unrefined) Rust type of x is S<&'?5 i32>, before the assignment, we identify a substitution that maps the region '?10 to '?5. After applying this substitution, the type of the place x is updated accordingly. This ensures that the lifetimes in the assigned type are consistent with those expected by the placeโs original type definition.

pub(crate) fn move_place( &mut self, rcx: &mut RefineCtxt<'_>, infcx: &InferCtxt<'_, '_, '_>, place: &Place, ) -> Result<Ty, CheckerErrKind>

pub(crate) fn unpack(&mut self, rcx: &mut RefineCtxt<'_>, check_overflow: bool)

pub(crate) fn unblock( &mut self, rcx: &mut RefineCtxt<'_>, place: &Place, check_overflow: bool, )

pub(crate) fn check_goto( self, rcx: &mut RefineCtxt<'_>, infcx: &mut InferCtxt<'_, '_, '_>, bb_env: &BasicBlockEnv, target: BasicBlock, ) -> Result<(), CheckerErrKind>

pub(crate) fn fold( &mut self, rcx: &mut RefineCtxt<'_>, infcx: &mut InferCtxt<'_, '_, '_>, place: &Place, ) -> Result<(), CheckerErrKind>

pub(crate) fn unfold( &mut self, infcx: &InferCtxt<'_, '_, '_>, rcx: &mut RefineCtxt<'_>, place: &Place, checker_conf: CheckerConfig, ) -> Result<(), CheckerErrKind>

pub(crate) fn downcast( &mut self, infcx: &InferCtxt<'_, '_, '_>, rcx: &mut RefineCtxt<'_>, place: &Place, variant_idx: VariantIdx, checker_config: CheckerConfig, ) -> Result<(), CheckerErrKind>

pub fn replace_evars(&mut self, evars: &EVarSol)