Crate flux_refineck

Refinement type checking

Modules

checker 🔒

errors 🔒

ghost_statements 🔒

Ghost statements are statements that are not part of the original mir, but are added from information extracted from the compiler or some additional analysis.

invariants

primops 🔒

queue 🔒

type_env 🔒

Structs

CheckerConfig

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

call_error 🔒

check_fn

force_conv 🔒

report_errors 🔒

ret_error 🔒