Crate flux_refineck

Refinement type checking

Modules

checker 🔒

constraint_gen 🔒

errors 🔒

fixpoint_encoding 🔒

Encoding of the refinement tree into a fixpoint constraint.

ghost_statements 🔒

Ghost statements are statements that are not part of the original mir, but are added from information extracted from the compiler or some additional analysis.

invariants

queue 🔒

refine_tree 🔒

sigs 🔒

type_env 🔒

Macros

_define_btys

_sig

Structs

CheckerConfig

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

call_error 🔒

check_fn

report_errors 🔒

ret_error 🔒