Function flux_refineck::checker::init_region_gen

fn init_region_gen(body: &Body<'_>) -> IndexGen<RegionVid>

During borrow checking, rustc generates fresh region variable ids for each structurally different position in a type. For example, given a function

fn foo<'a, 'b>(x: &'a S<'a>, y: &'b u32)

rustc will generate variables ?2 and ?3 for the universal regions 'a and 'b (the variable ?0 correspond to 'static and ?1 to the implicit lifetime of the function body). Additionally, it will assign x type &‘?4 S<’?5>andytype&’?6 u32` (together with some constraints relating region variables).

The exact ids picked for 'a and 'b are not too relevant to us, the important part is the regions used in the types of x and y. To recover the correct regions, whenever there’s an assignment of a refinement type T to a variable with (unrefined) Rust type S, we match both types to infer a region substition. For this to work, we need to give a different variable id to every position in T. To avoid clashes, we need to use fresh ids, so we start enumerating from the last id generated by borrow checking.

The ids generated during refinement type checking are purely instrumental and they should never appear in a type bound in the environment.

Besides generating ids when checking a function’s body, we also need to generate fresh ids at function calls.