Struct flux_refineck::checker::Checker

pub(crate) struct Checker<'ck, 'tcx, M> {

    genv: &'ck GlobalEnv<'ck, 'tcx>,
    config: CheckerConfig,
    def_id: LocalDefId,
    generics: Generics,
    refparams: List<Expr>,
    body: &'ck Body<'tcx>,
    resume_ty: Option<Ty>,
    ghost_stmts: &'ck UnordMap<LocalDefId, GhostStatements>,
    output: Binder<FnOutput>,
    mode: &'ck mut M,
    snapshots: IndexVec<BasicBlock, Option<Snapshot>>,
    visited: BitSet<BasicBlock>,
    queue: WorkQueue<'ck>,

}

Fields
genv: &'ck GlobalEnv<'ck, 'tcx>config: CheckerConfigdef_id: LocalDefId

LocalDefId of the function-like item being checked.

generics: Generics

[Generics] of the function being checked.

refparams: List<Expr>

[Expr]s used to instantiate EarlyBinders for signature of function being checked

body: &'ck Body<'tcx>resume_ty: Option<Ty>

The type used for the resume argument of a generator.

ghost_stmts: &'ck UnordMap<LocalDefId, GhostStatements>output: Binder<FnOutput>mode: &'ck mut Msnapshots: IndexVec<BasicBlock, Option<Snapshot>>

A snapshot of the refinement context at the end of the basic block after applying the effects of the terminator.

visited: BitSet<BasicBlock>queue: WorkQueue<'ck>
Implementations

impl<'a, 'tcx> Checker<'a, 'tcx, ShapeMode>

pub(crate) fn run_in_shape_mode( genv: &GlobalEnv<'a, 'tcx>, def_id: LocalDefId, ghost_stmts: &'a UnordMap<LocalDefId, GhostStatements>, config: CheckerConfig ) -> Result<ShapeResult, CheckerError>

impl<'a, 'tcx> Checker<'a, 'tcx, RefineMode>

pub(crate) fn run_in_refine_mode( genv: &GlobalEnv<'a, 'tcx>, def_id: LocalDefId, ghost_stmts: &'a UnordMap<LocalDefId, GhostStatements>, bb_env_shapes: ShapeResult, config: CheckerConfig ) -> Result<(RefineTree, KVarStore), CheckerError>

impl<'a, 'tcx, M: Mode> Checker<'a, 'tcx, M>

fn run( genv: &'a GlobalEnv<'a, 'tcx>, refine_tree: RefineSubtree<'a>, def_id: LocalDefId, ghost_stmts: &'a UnordMap<LocalDefId, GhostStatements>, mode: &'a mut M, poly_sig: EarlyBinder<PolyFnSig>, refparams: Option<List<Expr>>, config: CheckerConfig ) -> Result<(), CheckerError>

fn init<'b>( rcx: &mut RefineCtxt<'_>, body: &'b Body<'_>, fn_sig: &FnSig, config: CheckerConfig ) -> TypeEnv<'b>

fn check_basic_block( &mut self, rcx: RefineCtxt<'_>, env: TypeEnv<'_>, bb: BasicBlock ) -> Result<(), CheckerError>

fn check_assign_ty( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, place: &Place, ty: Ty, source_info: SourceInfo ) -> Result<(), CheckerError>

fn check_statement( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, stmt: &Statement ) -> Result<(), CheckerError>

fn is_exit_block(&self, bb: BasicBlock) -> bool

fn check_terminator( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, terminator: &Terminator<'tcx>, last_stmt_span: Option<Span> ) -> Result<Vec<(BasicBlock, Guard)>, CheckerError>

For check_terminator, the output Vec<BasicBlock, Guard> denotes,

BasicBlock “successors” of the current terminator, and

Guard are extra control information from, e.g. the SwitchInt (or Assert) you can assume when checking the correspondnig successor.

fn check_call( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, terminator_span: Span, did: Option<DefId>, fn_sig: EarlyBinder<PolyFnSig>, generic_args: &[GenericArg], actuals: &[Ty] ) -> Result<Ty, CheckerError>

fn check_oblig_generator_pred( &mut self, rcx: &mut RefineCtxt<'_>, snapshot: &Snapshot, gen_pred: GeneratorObligPredicate ) -> Result<(), CheckerError>

fn check_oblig_fn_trait_pred( &mut self, rcx: &mut RefineCtxt<'_>, snapshot: &Snapshot, fn_trait_pred: FnTraitPredicate ) -> Result<(), CheckerError>

fn check_closure_obligs( &mut self, rcx: &mut RefineCtxt<'_>, obligs: Obligations ) -> Result<(), CheckerError>

This checks obligations related to closures & generators; the remainder are directly checked in check_fn_call

fn check_assert( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, terminator_span: Span, cond: &Operand, expected: bool, msg: &AssertKind ) -> Result<Guard, CheckerError>

fn check_if(discr_ty: &Ty, targets: &SwitchTargets) -> Vec<(BasicBlock, Guard)>

fn check_match( discr_ty: &Ty, targets: &SwitchTargets ) -> Vec<(BasicBlock, Guard)>

fn check_successors( &mut self, rcx: RefineCtxt<'_>, env: TypeEnv<'_>, from: BasicBlock, terminator_span: Span, successors: Vec<(BasicBlock, Guard)> ) -> Result<(), CheckerError>

fn check_goto( &mut self, rcx: RefineCtxt<'_>, env: TypeEnv<'_>, from: BasicBlock, span: Span, target: BasicBlock ) -> Result<(), CheckerError>

fn check_rvalue( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, stmt_span: Span, rvalue: &Rvalue ) -> Result<Ty, CheckerError>

fn check_aggregate_operands( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, stmt_span: Span, args: &[Operand] ) -> Result<Result<Vec<Interned<TyS>>, CheckerError>, CheckerError>

fn check_len( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, source_span: Span, place: &Place ) -> Result<Ty, CheckerError>

fn check_binary_op( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, source_span: Span, bin_op: BinOp, op1: &Operand, op2: &Operand ) -> Result<Ty, CheckerError>

fn check_unary_op( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, source_span: Span, un_op: UnOp, op: &Operand ) -> Result<Ty, CheckerError>

fn check_cast( &self, kind: CastKind, from: &Ty, to: &Ty ) -> Result<Ty, CheckerError>

fn check_operands( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, source_span: Span, operands: &[Operand] ) -> Result<Vec<Ty>, CheckerError>

fn check_operand( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, source_span: Span, operand: &Operand ) -> Result<Ty, CheckerError>

fn check_constant(&mut self, c: &Constant) -> Result<Ty, CheckerError>

fn check_ghost_statements_at( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, point: Point, span: Span ) -> Result<(), CheckerError>

fn check_ghost_statement( &mut self, rcx: &mut RefineCtxt<'_>, env: &mut TypeEnv<'_>, stmt: &GhostStatement, span: Span ) -> Result<(), CheckerError>

fn constr_gen( &mut self, rcx: &RefineCtxt<'_>, span: Span ) -> ConstrGen<'_, 'tcx>

fn snapshot_at_dominator(&self, bb: BasicBlock) -> &Snapshot

fn dominators(&self) -> &'a Dominators<BasicBlock>

fn ghost_stmts(&self) -> &'a GhostStatements