Module flux_refineck::fixpoint_encoding

Encoding of the refinement tree into a fixpoint constraint.

Modules

fixpoint

Structs

ConstInfo 🔒

Env 🔒

Environment used to map between rty::Var and fixpoint::LocalVar. This only supports mapping of rty::Var::LateBound and rty::Var::Free.

ExprEncodingCtxt 🔒

FixpointCtxt

FixpointKVar 🔒

KVarDecl 🔒

KVarEncodingCtxt 🔒

KVarStore

SortStore 🔒

Keep track of all the data sorts that we need to define in fixpoint to encode the constraint. Currently, we encode all aggregate sorts as a tuple.

TagIdx

Enums

KVarEncoding

How an rty::KVar is encoded in the fixpoint constraint

Key 🔒

Functions

alias_reft_sort 🔒

This function returns a very polymorphic sort for the UIF encoding an associated refinement. This is ok, as well-formedness in previous phases will ensure the function is always instantiated with the same sorts. However, the proper thing is to compute the actual mono-sort at which the associated refinement is being used see GlobalEnv::sort_of_alias_reft but that is a bit tedious as its done using the fhir (not rty). Alternatively, we might stash the computed mono-sort in the rty::AliasReft during conv?

fixpoint_const_map 🔒

func_sort_to_fixpoint 🔒

sort_to_fixpoint

stitch

tuple_sort_name 🔒

Type Aliases

Bindings

An alias for additional bindings introduced when ANF-ing index expressions in the course of encoding into fixpoint.

ConstMap 🔒

KVidMap 🔒

PredSpans 🔒

An alias for a list of predicate (conjuncts) and their spans, used to give localized errors when refine checking fails.