Module flux_refineck::fixpoint_encoding

Encoding of the refinement tree into a fixpoint constraint.

Modules

fixpoint

Structs

ConstInfo 🔒

Env 🔒

Environment used to map between rty::Var and fixpoint::LocalVar. This only supports mapping of rty::Var::Bound and rty::Var::Free.

ExprEncodingCtxt 🔒

FixpointCtxt

FixpointKVar 🔒

KVarDecl 🔒

KVarEncodingCtxt 🔒

During encoding into fixpoint we generate multiple fxpoint kvars per kvar in flux. A KVarEncodingCtxt is used to keep track of the state needed for this.

KVarStore

SortStore 🔒

Keep track of all the data sorts that we need to define in fixpoint to encode the constraint. Currently, we encode all aggregate sorts as a tuple.

TagIdx

Enums

KVarEncoding

How an rty::KVar is encoded in the fixpoint constraint

Key 🔒

Functions

bv_size_to_fixpoint 🔒

func_sort_to_fixpoint 🔒

mk_implies 🔒

sort_to_fixpoint

tuple_sort_name 🔒

Type Aliases

ConstMap 🔒