Module flux_refineck::ghost_statements

Ghost statements are statements that are not part of the original mir, but are added from information extracted from the compiler or some additional analysis.

Modules

fold_unfold ğŸ”’

points_to ğŸ”’

This module implements a points-to analysis for mutable references.

Structs

GhostStatements ğŸ”’

StatementsAt ğŸ”’

Enums

GhostStatement ğŸ”’

Point ğŸ”’

A point in the control flow graph where ghost statements can be inserted.

Functions

all_nested_bodies ğŸ”’

compute_ghost_statements ğŸ”’

Type Aliases

EdgeMap ğŸ”’

LocationMap ğŸ”’