flux_refineck::ghost_statements::points_to

Struct PointsToAnalysis

struct PointsToAnalysis<'a> {
    fn_sig: Option<&'a EarlyBinder<PolyFnSig>>,
    map: &'a Map,
}

This implement a points to analysis for mutable references over a [FlatSet]. The analysis is a may analysis. If you want to know if a reference definitively points to a location you have to combine it with the result of a definitely initialized analysis. See module level documentation for more details.

Fields
fn_sig: Option<&'a EarlyBinder<PolyFnSig>>map: &'a Map
Implementations

impl<'a> PointsToAnalysis<'a>

fn new(map: &'a Map, fn_sig: Option<&'a EarlyBinder<PolyFnSig>>) -> Self

fn handle_statement(&self, statement: &Statement<'_>, state: &mut State)

fn handle_assign( &self, target: Place<'_>, rvalue: &Rvalue<'_>, state: &mut State, )

fn handle_ref(&self, place: &Place<'_>, state: &State) -> FlatSet<Loc>

fn handle_operand(&self, operand: &Operand<'_>) -> Option<PlaceIndex>

fn handle_terminator<'mir, 'tcx>( &self, terminator: &'mir Terminator<'tcx>, state: &mut State, ) -> TerminatorEdges<'mir, 'tcx>

The effect of a successful function call return should not be applied here, see [Analysis::apply_primary_terminator_effect].

fn handle_call_return( &self, return_places: CallReturnPlaces<'_, '_>, state: &mut State, )