flux_refineck::type_env::place_ty

Function downcast_struct

fn downcast_struct(
    infcx: &mut InferCtxt<'_, '_, '_>,
    adt: &AdtDef,
    args: &[GenericArg],
    idx: &Expr,
    span: Span,
) -> InferResult<Vec<Ty>>

downcast on struct works as follows Given a struct definition struct S<A..>[{i, …}] { fld : T, …} and a * “place” x: S<t..>[{e,..}] the downcast returns a vector of ty for each fld of x where * x.fld : T[A := t ..][i := e...] i.e. by substituting the type and value indices using the types and values from x.