Struct flux_refineck::type_env::TypeEnv

pub struct TypeEnv<'a> {
    bindings: PlacesTree,
    local_decls: &'a LocalDecls,
}

Fields
bindings: PlacesTreelocal_decls: &'a LocalDecls
Implementations

impl TypeEnv<'_>

pub fn new(local_decls: &LocalDecls) -> TypeEnv<'_>

pub fn alloc_universal_loc(&mut self, loc: Loc, place: Place, ty: Ty)

pub fn alloc_with_ty(&mut self, local: Local, ty: Ty)

pub fn alloc(&mut self, local: Local)

pub(crate) fn into_infer( self, scope: Scope ) -> Result<BasicBlockEnvShape, CheckerErrKind>

pub(crate) fn lookup_place( &mut self, genv: GlobalEnv<'_, '_>, rcx: &mut RefineCtxt<'_>, place: &Place ) -> Result<Ty, CheckerErrKind>

pub(crate) fn get(&self, path: &Path) -> Ty

pub fn update_path(&mut self, path: &Path, new_ty: Ty)

pub(crate) fn borrow( &mut self, genv: GlobalEnv<'_, '_>, rcx: &mut RefineCtxt<'_>, re: Region, mutbl: Mutability, place: &Place ) -> Result<Ty, CheckerErrKind>

When checking a borrow in the right hand side of an assignment x = &'?n p, we use the annotated region '?n in the type of the result. This region will only be used temporarily and then replaced by the region in the type of the x after the assignment.

pub(crate) fn ptr_to_borrow( &mut self, rcx: &mut RefineCtxt<'_>, gen: &mut ConstrGen<'_, '_, '_>, place: &Place ) -> Result<(), CheckerErrKind>

Converts a pointer ptr(mut, l) to a borrow &mut T for a type T that needs to be inferred. For example, given the environment

x: i32[a], p: ptr(mut, x)

Converting the pointer to a borrow will produce

x: †i32{v: $k(v)}, p: &mut i32{v: $k(v)}

together with a constraint

i32[a] <: i32{v: $k(v)}

pub(crate) fn assign( &mut self, rcx: &mut RefineCtxt<'_>, gen: &mut ConstrGen<'_, '_, '_>, place: &Place, new_ty: Ty ) -> Result<(), CheckerErrKind>

pub(crate) fn move_place( &mut self, genv: GlobalEnv<'_, '_>, rcx: &mut RefineCtxt<'_>, place: &Place ) -> Result<Ty, CheckerErrKind>

pub(crate) fn unpack(&mut self, rcx: &mut RefineCtxt<'_>, check_overflow: bool)

pub(crate) fn unblock( &mut self, rcx: &mut RefineCtxt<'_>, place: &Place, check_overflow: bool )

pub(crate) fn block_with( &mut self, genv: GlobalEnv<'_, '_>, path: &Path, new_ty: Ty ) -> Result<Ty, CheckerErrKind>

pub(crate) fn check_goto( self, rcx: &mut RefineCtxt<'_>, gen: &mut ConstrGen<'_, '_, '_>, bb_env: &BasicBlockEnv, target: BasicBlock ) -> Result<(), CheckerErrKind>

pub(crate) fn fold( &mut self, rcx: &mut RefineCtxt<'_>, gen: &mut ConstrGen<'_, '_, '_>, place: &Place ) -> Result<(), CheckerErrKind>

pub(crate) fn unfold( &mut self, genv: GlobalEnv<'_, '_>, rcx: &mut RefineCtxt<'_>, place: &Place, checker_conf: CheckerConfig ) -> Result<(), CheckerErrKind>

pub(crate) fn downcast( &mut self, genv: GlobalEnv<'_, '_>, rcx: &mut RefineCtxt<'_>, place: &Place, variant_idx: VariantIdx, checker_config: CheckerConfig ) -> Result<(), CheckerErrKind>

pub fn replace_evars(&mut self, evars: &EVarSol)